Adds a context provider to handle multiselection
